Move the export view with the rest of the views.
